% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{bionames}
\alias{bionames}
\title{Print Bioclimatic Variable Names}
\usage{
bionames(bios = 1:35)
}
\arguments{
\item{bios}{Numeric vector indicating the indices of bioclimatic variables to print.
Default is 1:35, which prints all variable names.}
}
\value{
None. Prints the names of the selected bioclimatic variables to the console.
}
\description{
This function prints the names of bioclimatic variables based on the specified indices.
}
\examples{
bionames()           # Print all bioclimatic variable names
bionames(c(1, 5, 12)) # Print names for variables 1, 5, and 12
}
