% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_periodic.R
\name{custom_periodic}
\alias{custom_periodic}
\title{Calculate custom version of periodic bioclims}
\usage{
custom_periodic(
  target_var,
  target_agg = "mean",
  period_var,
  period_agg,
  period = 3,
  circular = FALSE,
  checkNA = TRUE,
  stopNA = TRUE,
  out_name = NA
)
}
\arguments{
\item{target_var}{Climate variable to summarize over defined periods}

\item{target_agg}{Method to aggregate target variable}

\item{period_var}{Climate variable used to define the periods (e.g. wettest period is defined by precipitation)}

\item{period_agg}{Method to select target period (e.g. 'max' for wettest period or 'min' or driest)}

\item{period}{numeric. Length of period to summarize data (e.g., quarters,
semesters). If using monthly data, a quarter (3-months) will be used.}

\item{circular}{TRUE/FALSE. Calculate periods that include first and last
units. For example, if using mean monthly data and quarters, circular=TRUE will also
calculate Nov-Dec-Jan and Dec-Jan-Feb.}

\item{out_name}{Optional, variable name in output raster}
}
\description{
Calculate custom version of periodic bioclims
}
\details{
For 'target_agg' the default value is 'mean'. This is typically correct, except when aggregating precipitation, 
then target_agg should be 'sum.'
}
