% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sum_period}
\alias{sum_period}
\title{Calculate sum of climate raster per period}
\usage{
sum_period(x, period, circular)
}
\arguments{
\item{x}{spatRaster}

\item{period}{Length of period. Default is three. If you are using months. It will be a quarter.}

\item{circular}{logical Include first month/weeks?}
}
\description{
Calculate sum of climate raster per period
}
