% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ventana}
\alias{ventana}
\title{Windows}
\usage{
ventana(x, period, circular)
}
\arguments{
\item{x}{spatRaster}

\item{period}{Length of period. Default is three. If you are using months. It will be a quarter.}

\item{circular}{logical Include first month/weeks?}
}
\description{
Windows
}
