% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clima.R
\name{clima}
\alias{clima}
\title{Get several bioclimatic variables}
\usage{
clima(
  bios,
  tmin = NULL,
  tmax = NULL,
  tavg = NULL,
  prcp = NULL,
  period = 3,
  circular = FALSE,
  checkNA = TRUE,
  stopNA = TRUE
)
}
\arguments{
\item{bios}{Numeric vector with bios number}

\item{tmin}{spatRaster with minimum temperature}

\item{tmax}{spatRaster with maximum temperature}

\item{tavg}{spatRaster with average temperature. It could be NULL if tmin and
tmax are provided}

\item{prcp}{spatRaster with precipitation}

\item{period}{numeric. Length of period to summarize data (e.g., quarters,
semesters). If using monthly data, a quarter (3-months) will be use to
calculated Bios 08, 09, 10, 11, 16, 17, 18, 19,}

\item{circular}{TRUE/FALSE. Calculated periods that include first and last
units. For example, if using mean monthly data and quarters, circular=TRUE will also
calculate Nov-Dec-Jan and Dec-Jan-Feb.}
}
\description{
Get several bioclimatic variables
}
